% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extJSLibs.R
\name{extJSLibs}
\alias{extJSLibs}
\title{Names and versions of external JavaScript libraries.}
\usage{
extJSLibs()
}
\value{
A named character vector with the versions of the \code{geoarrow}, \code{arrow},
\code{parquet-wasm}, \code{geoparquet-wasm} & \code{flatgeobuf-wasm} JavaScript libraries
shipped with this package.
}
\description{
Names and versions of the external JavaScript libraries used in
\code{geoarrowWidget}.
}
\details{
See e.g.
\itemize{
\item \url{https://cdn.jsdelivr.net/npm/apache-arrow/package.json},
\item \url{https://cdn.jsdelivr.net/npm/@geoarrow/geoarrow-js/package.json},
\item \url{https://cdn.jsdelivr.net/npm/parquet-wasm/package.json},
\item \url{https://cdn.jsdelivr.net/npm/@geoarrow/geoparquet-wasm/package.json}
}

for more details on the JavaScript depencencies.
}
\examples{
  extJSLibs()

}
