% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachGeoarrowDependencies.R
\name{attachGeoarrowDependencies}
\alias{attachGeoarrowDependencies}
\title{Attach \code{arrow} and \code{geoarrow} JavaScript dependencies to a widget.}
\usage{
attachGeoarrowDependencies(x)
}
\arguments{
\item{x}{A widget created with \code{\link[htmlwidgets]{createWidget}}.}
}
\description{
Pipe-friendly function to attach \code{arrow} and \code{geoarrow} JavaScript dependencies
to a widget created with \code{\link[htmlwidgets]{createWidget}}.
}
\examples{
library(listviewer)

jsonedit(
  list("Just some dummy text")
  , elementId = "lv-example"
) |>
  attachGeoarrowDependencies()

## open the resulting page in the browser and inspect the page source, e.g.
## by pressing <Ctrl + u>. You should see two lines like:

# <script src="lib/apache-arrow-js-19.0.1/Arrow.es2015.min.js"></script>
# <script src="lib/geoarrow-js-0.3.2/geoarrow.umd.min.js"></script>

## !Version numbers of the JavaScript dependencies may differ!

}
