% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachGeoArrowDependencies.R
\name{attachGeoarrowDependencies}
\alias{attachGeoarrowDependencies}
\alias{attachGeoarrowDependency}
\alias{attachArrowDependency}
\title{Attach \code{Arrow} and \code{Geoarrow} JavaScript dependencies to a widget.}
\usage{
attachGeoarrowDependencies(x)

attachGeoarrowDependency(x)

attachArrowDependency(x)
}
\arguments{
\item{x}{A widget created with \code{\link[htmlwidgets]{createWidget}}.}
}
\description{
Pipe-friendly functions to attach \code{Arrow} and \code{Geoarrow} JavaScript dependencies
to a widget created with \code{\link[htmlwidgets]{createWidget}}.
}
\examples{
library(listviewer)

wgt = jsonedit(
  list("Just some dummy text")
  , elementId = "lv-example"
)
attachGeoarrowDependencies(wgt)

## open the resulting page in the browser and inspect the page source, e.g.
## by pressing <Ctrl + u>. You should see two lines like:

# <script src="lib/apache-arrow-js-19.0.1/Arrow.es2015.min.js"></script>
# <script src="lib/geoarrow-js-0.3.2/geoarrow.umd.min.js"></script>

## !Version numbers of the JavaScript dependencies may differ!

}
