% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extJSLibs.R
\name{extJSLibs}
\alias{extJSLibs}
\title{Names and versions of external JavaScript libraries.}
\usage{
extJSLibs()
}
\value{
A named character vector with the versions of the \code{GeoArrow}, \code{Arrow}
and \code{Parquet-WASM} JavaScript libraries shipped with this package.
}
\description{
Names and versions of the external JavaScript libraries used in
\code{geoarrowWidget}.
}
\details{
See e.g. \url{https://cdn.jsdelivr.net/npm/apache-arrow/package.json}
or \url{https://cdn.jsdelivr.net/npm/@geoarrow/geoarrow-js/package.json} or
\url{https://cdn.jsdelivr.net/npm/parquet-wasm/package.json} for
more details on the JavaScript depencencies.
}
\examples{
  extJSLibs()

}
