% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extJSLibs.R
\name{extJSLibs}
\alias{extJSLibs}
\title{Print names and versions of external JavaScript libraries.}
\usage{
extJSLibs()
}
\description{
Prints names and versions of the external JavaScript libraries used in
\code{geoarrowWidget}.
}
\details{
See e.g. \url{https://cdn.jsdelivr.net/npm/apache-arrow/package.json}
or \url{https://cdn.jsdelivr.net/npm/@geoarrow/geoarrow-js/package.json} for
more details on the js depencencies.
}
