% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoarrowWidget-package.R
\docType{package}
\name{geoarrowWidget-package}
\alias{geoarrowWidget}
\alias{geoarrowWidget-package}
\title{geoarrowWidget: Attach Geoarrow Data to a Widget}
\description{
Provides functions and necessary JavaScript bindings to quickly transfer geospatial data from R memory to the browser for e.g. interactive visualisation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-spatial/geoarrowWidget}
  \item \url{https://r-spatial.github.io/geoarrowWidget/}
  \item Report bugs at \url{https://github.com/r-spatial/geoarrowWidget/issues}
}

}
\author{
\strong{Maintainer}: Tim Appelhans \email{tim.appelhans@gmail.com}

}
\keyword{internal}
