% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoarrowWidget-package.R
\docType{package}
\name{geoarrowWidget-package}
\alias{geoarrowWidget}
\alias{geoarrowWidget-package}
\title{geoarrowWidget: Attach '(Geo)Arrow' and/or '(Geo)Parquet' Data to a Widget}
\description{
Provides functions and necessary 'JavaScript' bindings to quickly transfer spatial data from R memory or remote URLs to the browser for use in interactive 'HTML' widgets created with the 'htmlwidgets' R package. Leverages 'GeoArrow' (\url{https://geoarrow.org/}) data representation for data stored in local R memory which is generally faster than traditional 'GeoJSON' by minimising the amount of copy, serialization and deserialization steps necessary for the data transfer. Furthermore, provides functionality and 'JavaScript' bindings to consume 'GeoParquet' (\url{https://geoparquet.org/}) files from remote URLs in the browser.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-spatial/geoarrowWidget}
  \item \url{https://r-spatial.github.io/geoarrowWidget/}
  \item Report bugs at \url{https://github.com/r-spatial/geoarrowWidget/issues}
}

}
\author{
\strong{Maintainer}: Tim Appelhans \email{tim.appelhans@gmail.com} (\href{https://orcid.org/0000-0002-9824-2707}{ORCID})

}
\keyword{internal}
